Decal click detection

+++ The WildPixels/InputHandler.h
@@ -49,17 +49,17 @@ private:
 	InputHandler();
 	~InputHandler();
 
-	float x_mouse_pos;
-	float y_mouse_pos;
+	double x_mouse_pos;
+	double y_mouse_pos;
 
 	bool shouldQuit = false;
 
 public:
 
 	// These return the screen space coordinates of mouse position [-1, 1]
-	static float* GetMouseX();
+	static double* GetMouseX();
 	// These return the screen space coordinates of mouse position [-1, 1]
-	static float* GetMouseY();
+	static double* GetMouseY();
 
 	static InputHandler* instance;
 	static void Initialize();
