+++ The WildPixels/RenderAble.h
@@ -19,6 +19,6 @@ public:
 	const bool IsRender() const;
 	const bool IsClickAble() const;
 
-	virtual const bool Colliding(float* x, float* y) = 0;
+	virtual const bool Colliding(double* x, double* y) = 0;
 	virtual void Render() = 0;
 };
